Automated Video Chaptering Using Python/requirements-updated.txt
pandas==2.2.1
numpy==1.26.3
matplotlib==3.8.3
scikit-learn==1.4.1
google-api-python-client==2.123.0
youtube-transcript-api==0.6.2
python-dotenv==1.0.0
